% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_epoch}
\alias{as_epoch}
\title{Convert a value to Epoch Time}
\usage{
as_epoch(x)
}
\arguments{
\item{x}{object to be coerced}
}
\description{
This function takes an input and converts it to the Unix epoch which is the 
number of seconds that have elapsed since January 1, 1970 at Midnight UTC.
}
