% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_user_id}
\alias{check_user_id}
\title{Validate a User Id}
\usage{
check_user_id(user_id)
}
\arguments{
\item{user_id}{numeric; the unique id to identify an Instagram user which can be 
found in the \code{"pk"} field on a \code{user} object}
}
\description{
This function checks whether a supplied user_id fits the pattern of one, mostly 
to alert users when supplying a username instead of an id.
}
\examples{
# check with a real user id
bieber_user_id <- 6860189 # returned using ig_get_user_id("justinbieber")
check_user_id(bieber_user_id)

# this will return a warning that we need a User Id, not a Username
#check_user_id("justinbieber")
}
