% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_signed_body}
\alias{generate_signed_body}
\title{Generate a Signed Body}
\usage{
generate_signed_body(list_data)
}
\arguments{
\item{list_data}{\code{list}; a list of data to be converted to JSON}
}
\value{
character
}
\description{
This function generates signed json body when sending information to the API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\examples{
login_data <- list('phone_id' = "012345678901-99thisisatest99",
                   '_csrftoken' = "012345678901-99thisisatest99",
                   'username' = "testuser",
                   'guid' = "012345678901-99thisisatest99",
                   'device_id' = "012345678901-99thisisatest99",
                   'password' = "testpassword",
                   'login_attempt_count' = 0)
signed_body <- generate_signed_body(login_data)
}
\keyword{internal}
