% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_uuid}
\alias{generate_uuid}
\title{Generate a Unique Id}
\usage{
generate_uuid(keep_dashes = TRUE)
}
\arguments{
\item{keep_dashes}{logical; an indicator of whether to keep the symbol "-" in 
the generated Id}
}
\value{
character
}
\description{
This function generates universally unique ids with or without dashes
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\examples{
id_w_dashes <- generate_uuid()
id_wo_dashes <- generate_uuid(keep_dashes=FALSE)
}
\keyword{internal}
