% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{ig_auth_refresh}
\alias{ig_auth_refresh}
\title{Refresh an existing Authorized Instagram Token}
\usage{
ig_auth_refresh(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
a \code{Token2.0} object (an S3 class provided by \code{httr})
}
\description{
Force the current OAuth to refresh. This is only needed for times when you 
load the token from outside the current working directory, it is expired, and 
you're running in non-interactive mode.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
