% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{ig_following_recent_activity}
\alias{ig_following_recent_activity}
\title{Get news feed of recent activities for accounts you follow}
\usage{
ig_following_recent_activity(return_df = TRUE, verbose = FALSE)
}
\arguments{
\item{return_df}{logical; do you want to return the results as a \code{tbl_df} 
with one row per entity or as a list with one element per entity?}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function returns notifications regarding the people you follow, such as 
what posts they've liked or that they've started following other people.
}
\examples{
\donttest{
my_following_recent_activity <- ig_following_recent_activity()
}
}
