% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{ig_generic_GET}
\alias{ig_generic_GET}
\title{A Generic GET Function}
\usage{
ig_generic_GET(relative_endpoint, query = NULL, item_accessor = NULL,
  return_df = TRUE, paginate = FALSE, page_index = 1,
  max_pages = 10, verbose = FALSE)
}
\description{
This is a function that can be used for most all endpoints of the API that 
return items that can be filtered with a query and typically need to be paginated. 
This reduces the amount of redundant code inside each unique function that 
pings the endpoints.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
