% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{ig_generic_POST}
\alias{ig_generic_POST}
\title{A Generic POST Function}
\usage{
ig_generic_POST(relative_endpoint, body_as_list, return_df = FALSE,
  verbose = FALSE)
}
\description{
This is a function that can be used for most all endpoints of the API that 
accept a signed body of JSON data for a POST. This reduces the amount of redundant 
code inside each unique function that pings the endpoints.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
