% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_followers}
\alias{ig_get_followers}
\title{Returns a dataframe with information of followers of the account if it is public}
\usage{
ig_get_followers(user_ID, return_df = TRUE, verbose = FALSE)
}
\arguments{
\item{return_df}{When return_df=TRUE, this function will return the results in dataframe form, if return_df=FALSE it will return a list}

\item{verbose}{When verbose=FALSE, this function will run without printing the intermediary steps}
}
\description{
This function needs testing
}
