% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_following}
\alias{ig_get_following}
\title{Get all users that a user follows}
\usage{
ig_get_following(user_id, max_id = NULL, return_df = TRUE,
  paginate = TRUE, max_pages = 10, verbose = FALSE)
}
\arguments{
\item{user_id}{integer; the unique id to identify an Instagram user which can be 
found in the \code{"pk"} field on a \code{user} object}

\item{max_id}{integer; the unique id identifying the oldest post that you 
would want to retrieve in this function call}
}
\description{
This fuction returns all of the users that the specified user follows
}
\examples{
\dontrun{
this_user_id <- ig_get_user_id("r4scatUVA")
ig_get_following(this_user_id)
}
}
