% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_location_feed}
\alias{ig_get_location_feed}
\title{Return a dataframe of posts at the specified location}
\usage{
ig_get_location_feed(location_string, return_df = TRUE,
  verbose = FALSE, paginate = TRUE, max_pages = 10)
}
\arguments{
\item{location_string}{A string of characters: Syntax: unknown yet}

\item{return_df}{When return_df=TRUE, this function will return the results in dataframe form, if return_df=FALSE it will return a list}

\item{verbose}{When verbose=FALSE, this function will run without printing the intermediary steps}

\item{paginate}{When paginate=TRUE, this function allows you to return multiple page in instagram, currently each page has 18 entries}

\item{max_pages}{When max_pages = 10 & paginate=TRUE this function will return all the entries down to the last entry of the 10th page}
}
\description{
This function filters by location and returns all the posts that have the same location_string
}
