% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_media_id}
\alias{ig_get_media_id}
\title{Return the media ID from a user_id and post index}
\usage{
ig_get_media_id(user_id, post_index)
}
\arguments{
\item{username}{character; the username of the Instagram user, typically referenced 
by using the "@" symbol}
}
\description{
This function accepts a Instagram user_id and the number of the post that you want
and will return its MediaID, which is needed for other functions
}
\examples{
\dontrun{

}
}
