% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_saved_feed}
\alias{ig_get_saved_feed}
\title{ig_get_saved_feed}
\usage{
ig_get_saved_feed(return_df = TRUE, verbose = FALSE, paginate = TRUE,
  max_pages = 10)
}
\description{
NEEDS DOCUMENTATION!!!!
}
