% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ig_input_data_validation}
\alias{ig_input_data_validation}
\title{Validate the input for an operation}
\usage{
ig_input_data_validation(input_data, operation = "")
}
\description{
Validate the input for an operation
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
