% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{ig_rank_token}
\alias{ig_rank_token}
\title{Return rank_token resulting from Basic auth routine}
\usage{
ig_rank_token(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
character; a string of the rank_token element of the current authorized 
API session; otherwise NULL
}
\description{
Return rank_token resulting from Basic auth routine
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
