% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ig_search_tags}
\alias{ig_search_tags}
\title{ig_search_tags}
\usage{
ig_search_tags(tag, max_id = NULL, return_df = TRUE, paginate = TRUE,
  max_pages = 10, verbose = FALSE)
}
\description{
NEEDS DOCUMENTATION!!!!
}
