% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ig_search_username}
\alias{ig_search_username}
\title{Search for a user's profile by username}
\usage{
ig_search_username(username, return_df = FALSE, verbose = FALSE)
}
\arguments{
\item{username}{character; the username of the Instagram user, typically referenced 
by using the "@" symbol}

\item{return_df}{logical; do you want to return the results as a \code{tbl_df} 
with one row per entity or as a list with one element per entity?}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function returns the details of a user's profile based on the supplied 
username (omit the @ symbol)
}
