% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{ig_sync_features}
\alias{ig_sync_features}
\title{Synchronise experiments}
\usage{
ig_sync_features(features = getOption("Rinstapkg.experiments"),
  verbose = FALSE)
}
\arguments{
\item{features}{character; a long string with commas separating each of the 
experimental values to be set}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function performs a sync of the conditions of the app you are working with. 
The experiment conditions must need to be set prior to logging in using \link{ig_auth} by 
first using \code{options(Rinstapkg.experiments="...")}.
}
\note{
This is mainly for backend functionality during login.
}
