% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{items_to_tidy_df}
\alias{items_to_tidy_df}
\title{Convert JSON List of Items to as Data Frame}
\usage{
items_to_tidy_df(l)
}
\arguments{
\item{x}{\code{list}; a list object where each element is an item}
}
\value{
\code{tbl_df}
}
\description{
This function takes a list of items, potentially deeply nested, and returns 
a \code{tbl_df} with one row per item
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\examples{
list_dat <- list(list("name"="Item 1", 
                      "letters"=list("cap_letter"="A", 
                                     "lower_letter"="a")), 
                 list("name"="Item 2", 
                      "letters"=list("cap_letter"="B", 
                                     "lower_letter"="b"))) 
tidy_df_list <- items_to_tidy_df(list_dat)
}
\keyword{internal}
