% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{rank_token_available}
\alias{rank_token_available}
\title{Check rank_token availability}
\usage{
rank_token_available(verbose = TRUE)
}
\value{
logical
}
\description{
Check if a rank_token is available in \code{\link{Rinstapkg}}'s internal
\code{.state} environment.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
