% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rinstapkg.R
\docType{package}
\name{Rinstapkg}
\alias{Rinstapkg}
\alias{Rinstapkg-package}
\title{\code{Rinstapkg} package}
\description{
An R package connecting to the Instagram API using tidy principles
}
\details{
An implementation of the Instagram API using tidy principles. This 
includes functions to get feeds and users, but also perform basic in-app 
functionality such as liking, commenting, following, and blocking. Use of this 
package means that you will not use it to spam, harass, or perform other nefarious 
acts.

Additional material can be found in the 
\href{https://github.com/reportmort/Rinstapkg}{README} on GitHub
}
