% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Rinstapkg_state}
\alias{Rinstapkg_state}
\title{Return the package's .state environment variable}
\usage{
Rinstapkg_state()
}
\description{
Return the package's .state environment variable
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
