% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{token_available}
\alias{token_available}
\title{Check token availability}
\usage{
token_available(verbose = TRUE)
}
\value{
logical
}
\description{
Check if a token is available in \code{\link{Rinstapkg}}'s internal
\code{.state} environment.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
