% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{ig_comment}
\alias{ig_comment}
\title{Comment on a Post}
\usage{
ig_comment(media_id, comment_text, verbose = FALSE)
}
\arguments{
\item{media_id}{numeric; the unique id to identify a post which can be 
found in the \code{id}, not the \code{pk} field, of posts returned via many of 
the functions retrieving feeds.}

\item{comment_text}{character; the text that would be posted as a comment 
underneath the post}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function takes the \code{media_id} of a post as well as a \code{comment_text}
and leaves the comment on the post.
}
