% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_followers}
\alias{ig_get_followers}
\title{ig_get_followers}
\usage{
ig_get_followers(user_ID, return_df = TRUE, verbose = FALSE,
  paginate = TRUE, max_pages = 10)
}
\description{
NEEDS DOCUMENTATION!!!!
THIS WORKS!!!!!!
}
