% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_hashtag_feed}
\alias{ig_get_hashtag_feed}
\title{ig_get_hashtag_feed}
\usage{
ig_get_hashtag_feed(hashtag_string, return_df = TRUE, verbose = FALSE,
  paginate = TRUE, max_pages = 10)
}
\arguments{
\item{hashtag_string}{A string of characters: do not include the hashtag at the beginning}
}
\description{
This function filters by hashtags and returns all posts that have the save hashtag string
}
