% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_user_feed}
\alias{ig_get_user_feed}
\title{ig_get_user_feed}
\usage{
ig_get_user_feed(user_id, return_df = TRUE, verbose = FALSE,
  paginate = TRUE, max_pages = 10)
}
\arguments{
\item{user_id}{An integer}
}
\description{
This function uses the user_id to return a dataframe of information from that account's feed
}
