% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_user_id}
\alias{ig_get_user_id}
\title{Return User ID}
\usage{
ig_get_user_id(username, return_df = FALSE)
}
\arguments{
\item{username}{A character string}
}
\description{
This function uses and instagram username to return the userID that you will need for other functions
}
\examples{
{
ig_get_user_id("r4sctest")
ig_get_user_id("r4scatUVA")
}
}
