% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_user_id}
\alias{ig_get_user_id}
\title{Return the user ID from a username}
\usage{
ig_get_user_id(username)
}
\arguments{
\item{username}{character; the username of the Instagram user, typically referenced 
by using the "@" symbol}
}
\description{
This function accepts a Instagram username and returns their \code{user_id}, which is 
needed for other functions
}
\examples{
\dontrun{
ig_get_user_id("justinbieber")
}
}
