% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_user_id}
\alias{ig_get_user_id}
\title{Get a User Id}
\usage{
ig_get_user_id(username, verbose = FALSE)
}
\arguments{
\item{username}{character; the username of the Instagram user. Omit the "@" symbol 
that is typically used when referencing another user}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function accepts a Instagram username and returns their \code{user_id}, 
which is needed for other functions
}
\examples{
\donttest{
ig_get_user_id("justinbieber")
}
}
