% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ig_get_user_profile}
\alias{ig_get_user_profile}
\title{Get a User's Profile}
\usage{
ig_get_user_profile(username, verbose = FALSE)
}
\arguments{
\item{username}{character; the username of the Instagram user. Omit the "@" symbol 
that is typically used when referencing another user}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function returns the details of a user's profile based on the supplied 
username (omitting the @ symbol)
}
\examples{
\donttest{
# get Jusin Bieber's profile and to see how many followers he has
bieber_follower_cnt <- ig_get_user_profile("justinbieber")$follower_count
}
}
