% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{ig_my_timeline}
\alias{ig_my_timeline}
\title{ig_my_timeline}
\usage{
ig_my_timeline(max_id = NULL, min_timestamp = NULL, return_df = TRUE,
  paginate = TRUE, max_pages = 10, verbose = FALSE)
}
\description{
NEEDS DOCUMENTATION!!!!
}
