% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{ig_my_timeline}
\alias{ig_my_timeline}
\title{Get the authenticated user's timeline feed}
\usage{
ig_my_timeline(max_id = NULL, min_timestamp = NULL,
  ranked_content = TRUE, return_df = TRUE, paginate = TRUE,
  max_pages = 10, verbose = FALSE)
}
\arguments{
\item{max_id}{integer; the unique id identifying the oldest post that you 
would want to retrieve in this function call}

\item{min_timestamp}{integer, date, or datetime; a value identifying the oldest post by date 
that you would want to retrieve in this function call. Dates and various datetime 
objects will be converted into an integer representing time in epoch (seconds since 
January 1st, 1970) since that is what the API requires}

\item{ranked_content}{logical; do you want the feed content to be sorted by rank?}
}
\value{
\code{tbl_df} or \code{list}
}
\description{
This function returns data that would appear in the authenticated user's 
timeline feed.
}
\examples{
\dontrun{
my_timeline <- ig_my_timeline()
}
}
